+++ genrl/environments/vec_env/__init__.py
@@ -1,6 +1,6 @@
 from genrl.environments.vec_env.utils import RunningMeanStd  # noqa
-from genrl.environments.vec_env.vec_monitor import VecMonitor  # noqa
-from genrl.environments.vec_env.vec_normalize import VecNormalize  # noqa
+from genrl.environments.vec_env.monitor import VecMonitor  # noqa
+from genrl.environments.vec_env.normalize import VecNormalize  # noqa
 from genrl.environments.vec_env.vector_envs import (  # noqa
     SerialVecEnv,
     SubProcessVecEnv,
